tui: Add review file browser module
The TUI file-review browser module owns both the review loop and the file
chooser after the package initializer move.

The file chooser has separate responsibilities: listing reviewable paths,
filtering entries, marking files, and applying marked file actions. Keeping
that logic beside the review loop hides the boundary between opening a
review and choosing targets.

This commit addresses that by moving chooser state, reviewable-file
listing, marked-file actions, and the block-target prompt into
git_stage_batch.tui.file_review.file_browser while browser.py calls those
entry points.

The following tests commit will move direct file-browser coverage and
architecture guards to the new module.
